Bug fix for XBox controllers.
Hack to fix the select button on xbox controllers. This button triggers an event for the "Back" button instead of "Select" button on some versions of android, such as the one running on Oculus Quest. This change remaps the "Back" button press to the "Select" button press. The downside is if the user actually wants to press the back button they won't be able to do it.